Refactor the users can view reports tests
Instead of having multiple tests for the different types of
report, we have one test that tests the different types of
report are viewable in the place we expect them to be.
There's also no need to test the order of the reports
here, as it's taken care of in the
OrganisationReportsFetcher specs.