Set build and deploy jobs as distinct jobs
This seperates them from each other and makes it
clearer as to what job does what. When we add the
deploy job for the training environment, this will also
enable us to parallelalise the deployment to the prod and
training environments, speeding up the build process.

I have removed the opening line from the
`deploy-terraform` script, (which sets the
`TF_VAR_docker_image` environment variable to the
`DOCKER_TAG` environment variable, as we are now
setting this variable directly when we set up the
environment variables.