Use `a11y_action_link` to make an accessible link
As there is currently only one download on the page, it's
probably fine to just have the link as "Download", but as
we add more files, we need to be clearer about what
exactly the user will be downloading. The
`a11y_action_link` method appends some context in a
hidden span tag for screen readers.