Use the variance fetcher in the controller
This updates the variance controller to use the
fetcher introduced in the last commit. I've had to
update the comment specs here, becuase we now
only show activities with variance. Rather than go
through the somewhat complex method of creating
a forecast (which is covered in other specs), I've
stubbed the `variance_for_report_financial_quarter`
method to return a variance.