Remove role=navigation from reports tabs
This has the effect of invalidating the HTML code. Screen readers notify users when they come to a list, and tell them how many items are in a list. If you don't
mark up a list using proper semantic markup in a
hierarchy, list items cannot inform the listener that
they are listening to a list when no parent is indicating
the presence of a list and the type of list.