Add breadcrumbed concern
The way the `breadcrums_on_rails` gem works is to build up a stack
of breadcrumb components, usually intended to add up with each controller action
that goes one level deeper.

Our app uses different controllers for "deeper" levels of the same activity,
so we need to build up the stack more or less manually.