Make sure users are active even after login
Currently, users that are made inactive, but still
logged in can access parts of the site while their
session remains valid (i.e. they don't logout, or their
session doesn't time out). To get around this edge
case, when fetching the `current_user`, we limit our
query to only current users. If an active user is not
found, we invalidate the session and return an error
message