Helper method to get country list with percentages
Because we are now collecting multiple benefitting
countries, the IATI spec insists we add a percentage
attribute to specify how much of the funding goes to
each country. We've agreed that for now, it's enough
to show the split as equal (or near as we can), so I've
added a helper that returns the data about the
countries, as well as the percentage split.

Becuase the IATI spec insists the total MUST add up
to `100.0`, so, for example we can't have three
activities with 33.33% of the share. With this in mind,
if we can't divide equally, we pop the  remainder onto
the end of the last item.