Backfill benefitting_countries where possible
If there is a `recipient_country`, and optionally one or more `intended_beneficiaries`, we can populate `benefitting_countries` by concatenating all of them into one field.

If there is a `recipient_region`, we populate `benefitting_countries` with all the countries in that region.

However, this leaves us with several regions that do not have a list of countries defined.

These regions are:
```
[{"code"=>"289", "name"=>"South of Sahara, regional"},
 {"code"=>"298", "name"=>"Africa, regional"},
 {"code"=>"389", "name"=>"North & Central America, regional"},
 {"code"=>"498", "name"=>"America, regional"},
 {"code"=>"619", "name"=>"Central Asia, regional"},
 {"code"=>"679", "name"=>"South Asia, regional"},
 {"code"=>"798", "name"=>"Asia, regional"},
 {"code"=>"889", "name"=>"Oceania, regional"},
 {"code"=>"998", "name"=>"Developing countries, unspecified"}]
```

In these cases we cannot infer the benefitting countries, and we output a list of the affected activities.