Dig down from other subregions
If a subregion is a higher specificity, we need to use
the region to subregion mapping to get all the
countries within those subregions. There are still
some activities that don't get saved, because their
region is set to `Developing countries, unspecified`.
We need to make a decision around what to do with
this.