Use all geographical information for benefitting_countries
There are activities which have `geography=recipient_region` but also have `recipient_country` and/or `intended_beneficiaries`.

There are activities which have `geography=recipient_country` but also have `recipient_region`.

While the two geography options should have been mutually exclusive, the reality is that users have managed to fill in more data than that, and we don’t want to risk losing it.

Our choices were to only use the data corresponding to the geography - and risk discarding some data; use any and all data we can find in the records - and risk having the wrong data; run a report for any ambiguous cases - and risk delaying the process indefinitely.

Following a discussion with Meyric, we have chosen the second option.

The legacy fields will not be removed until BEIS are satisfied that the data is in the shape they want.