Breadcrumb trail for BEIS users
Because BEIS users can access all activities from all delivery partner organisations, the default breadcrumb trail scheme for activities now includes the path to the current or historical activities of the DP organisation.

Fund and programme activities are special cases: their organisation is the service owner, BEIS, but there is no index page to which a breadcrumb trail would apply.

We’ve chosen to include a link to the fund activity when looking at a programme, to be consistent with the existence of a link to the programme when looking at projects and 3rd party projects.