Set aria-controls value correctly
Previously, the `aria-controls` attribute of the button
was set to a value that didn't exist on the page. The
purpose of `aria-controls` is to show assistive
technology what element(s) are controlled by a given
element (https://www.w3.org/TR/wai-aria-1.1/#aria-controls).

In our case, this is the child activities that are shown
when clicking on the "expand" button. With this in
mind, I'm grabbing all the child activity elements, and
adding a space seperated list of the child IDs
(`aria-controls` accepts a list of IDs, as well as a
single ID). This tells assistive technology that clicking
this button will show/hide those elements.