Move all the activity tabs to the show controller
It's getting confusing having multiple controllers with one `show` action
for various different types of view of the same activity. This adds a new
service class called `Activity::Tab`, which accepts an activity, a current
user and a tab name. Assuming the tab name is correct we return a
template name and call the appropriate method, which sets various
instance variables. The `locals` method then sucks up all the instance
variables that have been set and then returns them as a hash. There is
also a `template` method that returns the template that we want to
render.

We then pass the two public methods to the `render` method of the
controller to tell the controller what template to render, and what local
variables the template has.