Only rebase the oldest PR set to automerge
Rather than needlessly rebasing all the PRs set to
automerge (which will then go out of date), I've
tweaked this to only get the oldest automerge PR
and rebase that one. Once that one is merged, it
will trigger a rebase on the second oldest one and
cascade down, saving us time and resources.