Switch out the aid type features for requests
This switches out the feature specs that test
specific behaviours when aid types are set for
much faster request specs. We have coverage for
much of this behaviour elsewhere in the test suite
and the feature specs are extremely slow.

I've also added some faster controller tests to
test when particular steps get skipped or shown,
which was previously tested in a bunch of
conditionals in the form helper
`fill_in_activity_form` method.