Simplify the programme level specs
This adds two "happy path" specs for adding
either GCRF or Newton activities, and checks if
the attributes have been set successfully. I've also
added individual methods to set variables on each
step, to begin to move away from the
`fill_in_activity_form`, which is unweildy and
loaded with conditionals.

I'm also deliberately not checking the contents of
the activity details page each time, as I'll add a
view spec for this later