Move some edit features to the view spec
There are a lot of feature specs that assert that
particular links and bits of text are visible on a
page. This is quicker to assert on a view spec. I've
also made some tweaks to the view spec to allow
the activities to be built, rather than creatives,
speeding up the spec further.