Do not fetch inactitve reports
Right now the OrganisationReportsFetcher is only used to fetch reports
for displaying to delivery partners.

Delivery partners should not be exposed to reports in the inactive
state so we chain onto the fetchers query with `no_inactive`.

I found the OrganisationReportsFetcher spec a little hard to follow, but
decided it was simply testing that the correct queries are run and we
get confidence from the fact we are using built in Rails scopes like
`not_approved` and `not_inactive` - which don't require testing.

I did add some additions the specs that I felt helped to document what
is happening in the methods.