Use the current_user instance variable
As all controllers inherit from the ApplicationController and that
controller includes the `Auth` concern, we can rely on the @current_user
instance variable being available in most circumstances.

Using an instance variable can make using view specs simpler as we can
`assign` them simply and easily.

Making this change lays the ground work for the next.

I've kept this commit separate to document my reasoning for this change.