Show an message when there are no active reports
The report status on the home page helps delivery partner users
understand what the shape of their reporting looks like.

When such a users has no report in a state the application considers
'active' (active, submitted, awaiting changes or in review) they do not
see any reports.

We want to show them a helpful message to help them understand why there
are no reports shown instead of a blank space.

We also want to link to the help centre to offer the user the
opportunity to get more in-depth help.

We use a view spec to test that the appropriate partial is rendered in
the expected states.