Use a page object for filling out activities
This improves on the activity form improvements
by using a version of the Page Object pattern
(https://thoughtbot.com/blog/better-acceptance-tests-with-page-objects).

This makes the helper easier to manage, as well
as allowing us to pass in our "template" activity
once and not having to pass it to every single
"micro-method". We also can have a
`created_activity` method, which returns the
activity that has been created by the form.