Set HistoricalEvent#trackable_type explicitly
In order for `HistoricalEvent#trackable_type` to be set to
`'Actual'` after:

```
HistoricalEvent.create(
  trackable: instance of Actual
)
```

we need to explicitly set the `trackable_type` of the STI
class. If not, the `HistoricalEvent#trackabletype` will be
the base class of the STI, i.e. `'Transaction'`.

The Rails docs cover this complexity in combining
polymorphic associations and Single-Table Inheritance:

https://railsdoc.github.io/classes/ActiveRecord/Associations/ClassMethods.html#module-ActiveRecord::Associations::ClassMethods-label-Polymorphic+Associations

However, the suggested technique of overriding
`HistoricalEvent#trackable_type=(class_name)` didn't work,
so here we use a 'before validation' hook to set the
attribute. This is the first callback available in the
Active Record lifecycle:

https://guides.rubyonrails.org/active_record_callbacks.html#available-callbacks)

It would be preferable to do the assignment correctly on
initialisation, so please improve this if you can!