Add timestamps to the budgets table
We could use the `timestamps` helper to add them the usual way, but the budgets table already has over 600 records without timestamps. The timestamps helper would create them with the `null: false` constraint, and then the migration would fail.

Rather than backfill the timestamps with made-up dates, we decided to allow them to be nullable.