Add some description of refund form validation errors
We're going to be making changes to the form to allow the Refund#comment
to be handled by a separate entity. In order to be confident that we are
continuing to handle the validation (and error-correction messaging) as
expected, we introduce a new scenario to describe how the user is directed
to correct errors if they submit the form without the necessary information.