Use a refund 'form object' for validation
Overview
--------

Just to step back from detail, this work is part of the plan
to recognise technically that Actuals, Refunds and
Adjustments are all types of Transaction.

We've decided to implement this using Single Table
Inheritance on the transactions table with the Transaction
model as the 'base' class implementing the polymorphism.

We've also (separately) recognised that we need to develop
our commenting ability so that comments can be applied to a
variety of entities in addition to Activities. These include
Refunds and Adjustments.

In this current work (changing `Refund` to use STI on
`Transaction`) we need to support the rule that refunds must
have a comment and we will be introducing a new
`FlexibleComment` entity which will allow us to transition to
a polymorphic commenting system. This is the planned
sequence of refactorings:

1. `Refund` uses `FlexibleComment`
2. `Adjustment` uses `FlexibleComment`
3. exisiting `Comment` entity migrated to `FlexibleComment`

This new form object
--------------------

We introduce this as a preliminary to extracting
`Refund#comment` to the coming `FlexibleComment`. The added
complexity of validating 2 models and representing their
validation error messages in the `refunds/_form` view can be
handled neatly in this way.