Add seperate code lists for regions and levels
This way, we can reduce duplication in the code
lists, and also allows us to refer directly to the
Developing countries, unspecified country in the
codelist, as opposed to it being in the code. I've
also added a codelist for the region levels, so we
can be super-clear about which level is which.

This also allows us to not have to rely on the
ordering of the regions to know which one is the
most specific.