Force word wrapping in error value cells
The contents of other cells is more predictable length-wise, but the error cell quoting the invalid value back at the user is user input data, and can be of unpredictable length.

In the case of concatenated country codes, this contents look like one long unbroken word, so we have to instruct the browser to break it up.