Add graduated to benefitting countries code list
Over time countries can 'graduate' off the ODA eligibility list.

The application may need to display graduated countries for historic
activities whilst only accepting non graduated countries for new
activities.

We want this modelling to be encoded in the code list if possible to
make identification by non developers simple and any changes as simple
as possible for developers when required.

Here we add the graduated countries to the code list and a flag to mark
the country as graduated or not.