The application shows countries from the code list
We want the countries code list to be the single place to derive country
information in the application.

Here, we update the activity view to use the `benefitting_countries`
code list to show country names. The application can show countries
regardless of if they have graduated or not.

As we have historical data that may contain countries from other
sources, we also handle any unexpected countries by rendering 'unknown
country', this will allow users to identify and deal with any
discrepancies by updating the activity or code list.