Extend BenefittingCountry
We are going to want to render the benefitting countries form from the
BenefittingCountry and BenefittingCountry::Region models.

To do so will require all regions at a given level and all countries in
a given region.

Here we add methods for each.

We also allow direct comparison of two Regions using `==` via the codes,
this allows us to perform more comparisons such as `include?` on Region
objects.