Describe the intended behaviour for recipient_region
The intended behaviour is already in place, but it needs to be explicitly described to prevent regressions.

The `recipient_region` field acts as a legacy field, in that it is not editable anymore, but it has to be shown unless the user specifically chooses to provide benefitting_countries.

This feature will become more visible once we run the data migration that undoes the previous data migration, which had populated the inferred countries from the given region.