Form uses benefitting countries
With the mechanics of benefitting countries now using the code list, we
want the web form to use it as well.

Users who are adding new activities or updating existing ones should
only be able to choose non graduated countries.

We have also added a link to the guidance in Zendesk to aid users to
respond to the question and changed the content to better reflect what
the controls do.

Whilst we know this UI is not perfect, it is good enough for the time
being to collect data and allow for user research and UI design to come
up with something that better meets users needs.