DP creates an adjustment via a form
- So that BEIS can generate accurate reports for SID
- As a DP
- I want to adjust "actual spend" in historic periods in
  line with corrections agreed with BEIS

FINANCIAL PERIOD

An adjustment:

- is automatically associated with a report in the "active"
  state, and
- must assigned to an explicit financial period

COMMENTS

It's agreed that it's essential that an adjustment includes
an explanation (or narrative) describing the background for
and rationale for the adjustment.

This is done with the new FlexibleComment class which is
going to replace the existing Comment class:

- `Refund` (done)
- `Adjustment` (here)
- `Activity` (to follow, "soon")

ADJUSTMENT TYPE

Adjustments must also be labelled as applying to either:

1. an Actual spend, or
2. a Refund

When reviewing spending BEIS use the "Download spending
breakdown as CSV file" link on a report page. This uses
`ActivitySpendingBreakdown` to output a CSV.
The downloaded file includes 3 columns per financial period
which concernAdjustment and Refund. To
represent spending, taking proper account of adjustments and
refunds, these columns should be calculated as follows:

i) *actual spend*: this should EXCLUDE refunds and any
   adjustments to refunds (but INCLUDE adjustments to actual
   spend)

ii) *actual refund:* this should INCLUDE any adjustments to
    refunds

iii) *actual net*: this is the total of actual spend LESS
     refunds. It should INCLUDE adjustments to both refunds
     and actuals

PRESENTATION

Similar to the approach taken with Refunds, we show
Adjustments in their own section of the Financials tab.

Each Adjustment listed on the Activity Financials tab links
to a 'show' view which provides full detail on the
individual Adjustment, including when it was posted, by
whom, the type of adjustment and the important narrative
comment. We also provide links to the associated Report and
Activity.