Include all regions at all levels
Some regions levels share the same name, but are in fact different
levels in the models.

For example:

Alegira (DZ) is in:

Region (level 1): Africa, regional
Subregion 1 (level 2): North of Sahara, regional
Subregion 2 (level 3): North of Sahara, regional

The simplest way to include this modelling is to make sure the region
levels are expressed in the code list as they are used, so here North of
Sahara, regional is both a level 2 and 3 region.