Fix region calculation
Whilst this work was happening we found a bug where the region
calculation fails under two conditions:

1. the activity has a graduated country in benefitting_contries
2. the activity has an unexpected country benefitting_countries

We now handle graduated countries in the BenefittingCountry model so
this is no longer an issue.

We felt there were two approaches to this:

1. alert the user in the interface that the region cannot be determined
2. remove the unexpected countries and calculate the region regardless

We decided to take the second option as we really do not know how many
activities actually have this issue and that number may be low.

We also know that the unexpected country will be listed in the
benefitting_countries data as 'unknown country', which is the source of
the problem and is a better place to help users understand and rectify
it.

If we learn that the number is high we can deal with that later.
This work extends BenefittingCountry to handle unexpected country codes
by ignoring them and continuing the calculation with the countries the
application does expect.