Split Region from BenefittingCountry
It was noted in a review that BenefittingCountry::Region was becoming a
first class citizen and we felt it should be split out in to its own
file.

We took the opportunity to rename it BenefittingRegion to keep the link
with the concept it is modelling as clear as we can.