Support multiple reports in the report export
This allows multiple reports to be exported using
the exporter. I've also added an `export_type`
argument to define the type of export, so we can
set the expected filename, as well as add a scope
to avoid N+1 queries. It's still really slow, but
this is something we can improve when we put
more functionality into the class itself.