Move export logic to the new export class
This moves the remaining logic from the
ExportActivityToCsv class and moves it to the
`Report::Export` class, adding a `Row` subclass
that generates the row data. Rather than checking
each time if the report has a financial quarter, we
filter all the reports without a quarter before
mapping through them.

There are also a bunch of other improvements that
should hopefully make this class easier to follow.