Handle multiple budgets for a financial year
Previously, we were only showing one row per
activity, meaning budgets were getting left out if
there was more than one budget for a financial
year. BEIS wanted any additional budget(s) to
show on a following line, to make totalling easy.

We handle this by looping through all the activities,
storing the budgets for the activity in an array.
Every time we come across a budget for a financial
year, we remove it from the array, and keep
outputting new rows until the budget array is
empty. For example, if we have a budget with the
values "200" and "20" for Financial year 2020 and
"400" for the Financial year 2021, the activity
would have two rows withe the following budget
data (with the activity data repeated on both
rows):

| 2020-2021 | 2021-2022 |
| 200.00    | 400.00    |
| 20.00     | 0.00      |