Rephrase specifications to reflect current reality
Remove the old spec referring to `recipient_country`, as this is now a legacy field, and all data from `recipient_country` and `intended_beneficiaries` has been migrated to `benefitting_countries`.

The only legacy field that needs to be respected is `recipient_region`, and only in the absence of any country-level data.