Introduce SpendingBreakdown::Export
The intention is for this to replace the existing
ActivitySpendingBreakdown class.

The previous class is based on the principle that the spending breakdown
is related to a report, BEIS finance users are not concerned with the
report time period, they are more interested in all the data the application
contains to date.

Here we introduce a class that takes a fund and optional organisation
and returns all the actual spend and refund values.

Now that we have the new Adjustment model, we also want to include the
correct type of adjustments in the calculations:

Actual spend total = Total actuals for an activity in a financial quarter  +
        Total actual adjustments in the same financial quarter

Refund total = Total refunds for an activity in a financial quarter +
        Total refund adjustments in the same financial quarter

Example:

Q1 2020-2021
-------------------------
Actual spend: 1,000
Actual spend: 10,000
-------------------------
Actual adjustment: -100
Actual adjustment: 50
-------------------------
Total: 10,950

The 'Net total' is the result of adding the Actual spend total to the
refund total.

This class follows our new pattern for csv exports of providing three
public methods to produce everything needed to export a csv.

This commit does not include the Forecasts in the file, that will
follow.

This commit does not include a way for users to intact with this class,
that will follow.