Allow adjustments to be included in overviews
This adds a named parameter to the
ActualOverview class to include adjustments. If
this is set to `true`, then the class includes the
adjustments in the Transaction scope, including
these in the totals for the quarter.

It’s arguable that we might want this to be the
default, but for now, we’ll keep this optional, in
the interests of keeping things simple.