Upgrade Ruby and JS dependencies
We also configure the test environment to not annotate the layout and
partials with the file name, we can only assume this was never working
as we have a test that now fails because of this.

It feels right that this is disabled in the test environment as it is
also disabled in production and we want parity between the two.