Fix Coherence and Impact GCRF strategic area code
When we added the code, we accidentally added `Clm` instead of `CIm`. (That’s a lowercase l-for-Lima instead of an uppercase i-for-India. An easy mistake to make, given that in many fonts they are almost identical.)

We need to run a data migration to fix the almost 300 existing activities with the wrong code.