Cache all transactions in an ActualOverview
This updates the ActualOverview to remove
the activity from the initializer. When returning
`all_quarters`, this now groups each
transaction by financial year, financial quarter
AND activity ID. This means that rather than
querying transactions on an activity by activity
basis for the export, we can load all of a
report's transactions into memory when we
start building the report, and pluck them from
the `AllQuarters` instance based on the
financial period and the activity.

I've also had to update the
`value_for_report_quarter` method to accept
an activity as an argument, so we can reuse
the relation in the class, but append the
query to search for a specific activity.