Add forecasts to SpendingBreakdown::Export
BEIS finance want to see all actual spending and all forecasted spending
on activities.

Here we add forecast values for the period after the last actual spend
values we have in the application.

Forecasts are already stored as a single total per activity per quarter,
so we do not need to sum them. We do have to interact with the forecast
history, this export will always include the latest values.

This commit sets up the operation and tests, next we will look to
refactor the code.