Speed up SpendingBreakdown::Export spec
As we want to exercise our database queries in the test, we have to
populate the database with all of the required entities, which is very
slow, especially when using `let` as the database entries are created
and destroyed for every example.

Here we create the database entries just once and use the
`DatabaseCleaner` gem (which we already have installed in the
application) to clear up after the spec file has been executed.

We also took some time to wrap the setup in some descriptive methods to
help document the setup and how it works along with the specs.

Spec time:

Before: ~30 seconds
After: ~3 seconds

Based on 1.4 GHz Quad-Core Intel Core i5 MacBook Pro