Add data migration to fix up over-eager backfilling of benefitting_countries
It seems we were over-eager in backfilling the new `Activity#benefitting_countries`
field, in that we added all of the countries in the legacy `recipient_region` even
when the legacy `recipient_country` field was present.

Example: Activity with RODA id 'GCRF-EP_C_MS-4_2018EP/T00343X/1':

   => {"recipient_region"=>"1027",
    "recipient_country"=>"UG",
    "intended_beneficiaries"=>nil,
    "requires_additional_benefitting_countries"=>false,
    "country_delivery_partners"=>[],
    "benefitting_countries"=>["UG", "BI", "KM", "DJ", "ER", "ET",
                              "KE", "MG", "MW", "MU", "MZ", "RW",
                              "SO", "SS", "SD", "TZ", "ZM", "ZW"]}

In this script, we find the approx 4,000 activities which are affected:

- legacy `recipient_region` is present, and
- legacy `recipient_country` is present, and
- there are no `intended_beneficiaries`, and
- more than 1 country is set in the new `benefitting_countries` field

and set the benefitting countries field to the legacy recipient_country value.
e.g. `["UG"]` in the example above.

It should produce output something like:

```
Fixing up 4130 activities...
-> there are now 0 activities still to fix
```