Update Node JS
We are currently building with Node 10, the current long term support
version is 14.18.0.

We've added a `.node-version` file to try and be explicit to developers
who want to ensure parity between their environment and the deployments.

Whilst this is not 100% reliable [1] it at least documents this
explicitly where developers will find it. Those that use a Node version
manager stand a good chance of it 'just working'.

[1] https://github.com/shadowspawn/node-version-usage