Show all geography related fields
The whole area of 'geography' has undergone significant changes.

We want to try and help users and not confuse them with all the changes
that have been happening.

Users cannot modify the legacy values, but whilst the data is
migrated to the new format it is helpful to clearly show the old values
for comparison with the new in order to validate the new values.

Here we show all the field regardless of whether they have a value or not.

We add a 'Legacy field: not editable' label to show users they cannot change the values.

We also remove the `requires additional benefitting countries` attribute
as it no longer serves any purpose.