Remove requires_additional_benefitting_countries
As we no longer show or need the value stored here, we can remove it and
any associated items.

The original purpose was to signal in the form wizard if more
benefitting countries are going to be added and it not skip that section
of the form. Now the app no longer collects
additional_benefitting_countries we can remove this column.

The actual country data is stored in `intended_beneficiaries` and we will not
remove that data until BEIS are confident to do so.