Report::Export can handle no comments
In the last commit we fixed the `comments_for_report` method on Activity,
here we test the new behaviour when used to create a report export,
previously this would have failed silently when there are no comments
and produced a csv file with no activities.

This would have been caused by comments_for_report returning nil and the
report export trying to map over the nil raising an error in the csv
creation that the user would never see.