Rename comment tables / models
This renames the FlexibleComment model to
Comment, and the existing Comment table to
LegacyComment. In a future commit, I'll be
adding a data migration to move all the
`LegacyComment`s to the new Comment
table. Once this is deployed and the data
migration is run, we can delete the
LegacyComment model and table.