Update report variance table comment links
Previously it was only possible to create one comment per activity, per
report. This restriction has been lifted by adding the ability to add
multiple comments on an activity from the Activity > Comments tab.

With this change in place the old add or view a comment against variance
no longer works as expected because there may be multiple comments.

We still want to keep the link between variance and adding a comment, as
we know that variance is a common scenario in which uses may leave a
comment, but not the only one.

Here we replace the add/view comment and view activity links on the
Report > Variance table with a simple link to:

- View and add comments if the use is authorised
- View comments otherwise

By doing so we retain the link between variance and an activity's
comments, whilst allowing any number of comments to be added to a single
activity during a report.

A user who wishes to add a comment from the variance table needs to take
an extra step to do so, but the relationship is more solid.

We used a view spec to test the new behaviour, which meant using
`without_partial_double_verification` to allow the controller to respond
to `policy` and return a stubbed policy to exercise both authorised and
unauthorised situations. This seems to be the only way to get a stubbed
policy into the view spec.