Data migration to move LegacyComments to Comments
This migrates the old, legacy comments to the
new polymorphic comment table.

There are some legacy comments that either
have no comment at all, or are related to an
activity that no longer exists, so we skip
creating a new comment for these, but keep a
record to report at the end.

Once this is in production, and run, we can
delete the old LegacyComments table and
model.