ChangeStateColumn on Report::Export
ChangeStateColumn encapsulates all of the knowledge to produce a
column of data for the change state in the report csv.

The motivation is to reduce database calls as much as possible by
memoising the results of queries.

Once instantiated with a report, it can return all the data for the
column including the header. The report is required as the report is the
context in which any change occurred.

The report also provides the fund and delivery partner organisation
used to scope the activities.

If an unexpected activity i.e. one that is not in the scope of the
report, is passed to ChangeStateColumn it will raise an ArgumentError.

Calling `state_of` with an activity will return one of the change
states:

- New: This activity was created in the report
- Changed: This activity's non-financial data has been modified in the
  report
- Unchanged: This activity was neither of the above in the report