Use ChangeStateColumn to populate the report csv
Here we use the new ChangeStateColumn class to populate the report csv.

Things are kept simple at this stage by inserting the change state
column after the activity's non-financial data.

Whilst this may not be the optimal location for the column, it is worth
remembering that users can manipulate the data in external tooling,
usually a spreadsheet application, and move columns to their own
preferred locations.

What bring value here is the ability for users to sort and filter the
data by the change state.

Doing so also means the changes to introduce the change state are
minimal and fit in with the existing data sources collected in the
report csv.