Introduce Commitment
The commitment is the total amount of funding BEIS commits at the outset
of an activity.

At this stage the primary motivation for storing the value is to output
it to the IATI XML to improve BEIS score on a forthcoming Publish What
You Fund score.

I decided to create a new table to store the commitment as it feels
right to keep all financial data out of the activity model, this also
gives flexibility later should it become apparent that multiple
commitments are required over the lifetime of an activity or we need to
store more detail about the commitment, the data range the commitment
covers for example.

There is a unique index on `activity_id` to prevent there ever being
more than one Commitment for an Activity.

To follow:

- importing commitments
- exposing commitments